ANTIZAVOZ/requirements.txt
python-dotenv==0.20.0
psycopg2-binary==2.8.6 #3.0.14
django-crispy-forms==1.14.0